python/requirements.txt
apigratis-sdk-python==1.1.7
python-dotenv==1.0.0